OCR/requirements.txt
google-cloud-vision==3.4.2
requests==2.31.0
Pillow==9.4.0